--- jax_imprl/__init__.py
@@ -1,2 +1,0 @@
-from .agents.primitives.agent import Agent, TransitionTuple
-from .agents.primitives.MLP import MLP
