+++ jax_imprl/agents/__init__.py
@@ -1,0 +1,11 @@
+import jax_imprl.agents
+
+from jax_imprl.agents.DDQN import DDQN
+from jax_imprl.agents.JAC import JointActorCritic as JAC
+
+
+def get_agent_class(algorithm):
+    try:
+        return getattr(jax_imprl.agents, algorithm)
+    except AttributeError:
+        raise NotImplementedError(f"The algorithm '{algorithm}' is not implemented.")
